big refactoring - split the fusion file into  smaller files for easier maintenance
